Tried Stacked, LGBM, and K Neighbor Models
Submitted the stacked model, lgbm model, and k neighbor model
predictions. K neighbor did the worst, is overfitting and not well
suited to this dataset. Stacked did well, but lgbm did better on its own
then in the stacked model so refinements have to be made to the other
models in the stacked model and how the models are stacked.